Implement the HDD star algorithm.
The original HDD approach defined the HDD star algorithm as a
fixed-point iteration of the hddmin method. In HDD*, such parts
of the input become removable, which are initially blocked by
semantic dependencies but what dependencies get removed in a
later iteration of the minimization. The patch implements this
variant into picireny, too.