Add option to build hidden tokens into the HDD tree
Most grammars skip whitespace and comments, or send them to a
hidden channel. Usually, these tokens are unnecessary both for HDD
and for the building (unparsing) of the test case, so they are not
built into the HDD tree. However, sometimes either comments or other
constructs/directives must be kept in the test case. This commit
adds an option to build even the hidden tokens into the tree.

In such cases, the existing unparsing strategy that heuristically
adds whitespace to the built test case is not optimal, so the
whitespace-insertion is disabled if hidden tokens are built in the
tree.