Rework ID, level, state concepts of HDD tree nodes - and how hddmin uses them
Now, there is no need for reassigning IDs to nodes on the current
level anymore, as IDs are given at node creation time. Level does
not need to be assigned to nodes at all, since nodes on a given
level can be collected just as easily without such an attribute.
Moreover, with proper planning, 'temporarily removed' state is not
needed, 'keep' and 'removed' states are enough to implement hddmin
(even for hddmin*). Finally, it is more logical to use a value for
the 'removed' state that evaluates to boolean false (0), and a
value for 'keep' that evaluates to true (1).