Test single-node levels for removal
DDMIN assumes that configurations smaller than 2 units cannot be
reduced further, and also assumes that empty configurations are
never failing. However, in HDD, this is not necessarily true.

(In HDD, removal actually means replacement with minimal valid
syntactic fragment. And a node can be a single element on a level
with smaller -- potentionally empty -- replacement than its
children on lower levels. If single-node levels are not tested for
removal -- i.e., replacement --, then removals on lower levels will
not be able to produce minimal results.)