Don't let empty nodes show up in ANTLR-built trees
EOF tokens and lambda-rules create empty nodes in the HDD tree:
EOF tokens have an empty string text, while lambda-rules have no
children (causing a non-token leaf in the tree). None of them add
anything useful to the tree but cause extra testing steps in the
algorithms.

This patch removes them in a post-processing step of the tree
builder.