Fix the replacement of the root of island subtrees.
Island subtrees are inserted into the tree by replacing the root
HDDToken by a HDDRule node which has the subtree as child. Until now,
only the name of this token node was copied into the new rule node and
not the replacement. As a consequence, the replacement of these
subtree roots were None, which is incorrect. The patch fixes this.