Allow unparsing to "preview" changes to the tree without actually applying those changes
This is achieved by adding an optional `transform` argument to
`unparse`, a function that takes a node and returns a potentially
changed one. When DD is executed, this argument is utilized and
nodes not in the tested config are not actually marked as removed,
but substituted by a copy that is marked as removed. Thus, state
change is only committed when the 1-minimal config is found.