Use auto-incremented IDs for tree nodes
This replaces the `id()`-based approach, which is neither
guaranteed to be stable across multiple runs nor easy to read (as
it is the memory address of the object, at least in CPython). IDs
generated by an auto-incremented counter don't have these problems.

Additionally, the tree dump format is extended to contain the node
IDs.