Handle tokens section in grammars
Identifiers in a `tokens { Token1, ..., TokenN }` section are
definitions of token names without an associated lexer rule. We
don't know anything about them, but they are added with a dummy
representation to the grammar tree that is used for minimal
replacement string calculation to avoid dead links (as they may be
referenced from other (parser) rules).

Fixes #16