Implement proper `__repr__` for the tree
This commit replaces the custom and - arguably - hard to read
string representation with a *mostly* valid Python expression
format.
(The generated expression passes some keyword arguments to the
constructors - id, state, children - that are not accepted (yet?).)

This commit also extends `log_tree` to print this tree
representation at `TRACE` log level. To get access to the `TRACE`
log level, this commit also replaces the `logging` package with
`picire.logging`.