Build-time generation of ANTLRv4Lexer and Parser
The ANTLRv4 lexer and parser are used to analyze the grammar of the
target language. However, the ANTLRv4 grammars, and the lexer and
parser generated from them are constant. Hence, building the lexer
and parser from the .g4 files at every reduction is redundant and a
waste of resources.

This patch moves the generation of the lexer and the parser from
run-time to build-time (using the build_antlr setuptools extension
provided by ANTLeRinator).

Co-authored-by: Renata Hodovan <reni@inf.u-szeged.hu>