Rename tox environment schema to jsonschema (#56)
Tox introduced a new command `schema` in 4.24.0, which conflicts
with the same environment name we used in `tox.ini`.

This happens because we call tox as `tox -e ENV`, relying on `run`
being the default command for tox. So, `tox run -e schema` would
also be an option, but that would require a lot of changes (if we
were to apply the change everywhere consistently). Thus, it is
easier to rename the environment instead.

Also: Update years in license