Enable subprocess patch for coverage measurement (#58)
From pytest-cov 7.0.0 (released on 2025-09-09), the support for
subprocess measurement has been dropped. This has caused drastic
loss of information (to the point that absolutely no coverage data
was collected and reported). The patch system of the underlying
coverage package (since 7.10) has a way to measure subprocesses
created in tests, but it needs to be enabled explicitly.